+++ src/main/java/com/flab/eattofit/exercise/application/sports/sports/dto/SportsCreateRequest.java
@@ -1,4 +1,4 @@
-package com.flab.eattofit.exercise.application.sports.dto;
+package com.flab.eattofit.exercise.application.sports.sports.dto;
 
 import jakarta.validation.constraints.NotEmpty;
