+++ src/main/java/com/flab/eattofit/exercise/domain/sports/sports/SportsRepository.java
@@ -1,4 +1,4 @@
-package com.flab.eattofit.exercise.domain.sports;
+package com.flab.eattofit.exercise.domain.sports.sports;
 
 public interface SportsRepository {
