+++ src/main/java/com/flab/eattofit/exercise/infrastructure/sports/sports/SportsJpaRepository.java
@@ -1,6 +1,6 @@
-package com.flab.eattofit.exercise.infrastructure.sports;
+package com.flab.eattofit.exercise.infrastructure.sports.sports;
 
-import com.flab.eattofit.exercise.domain.sports.Sports;
+import com.flab.eattofit.exercise.domain.sports.sports.Sports;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface SportsJpaRepository extends JpaRepository<Sports, Long> {
