+++ src/main/java/com/flab/eattofit/exercise/ui/sports/sports/dto/SportsCreateResponse.java
@@ -1,4 +1,4 @@
-package com.flab.eattofit.exercise.ui.sports.dto;
+package com.flab.eattofit.exercise.ui.sports.sports.dto;
 
 public record SportsCreateResponse(
         String name
