+++ src/test/java/com/flab/eattofit/exercise/application/sports/sports/SportsServiceTest.java
@@ -1,10 +1,10 @@
-package com.flab.eattofit.exercise.application.sports;
+package com.flab.eattofit.exercise.application.sports.sports;
 
-import com.flab.eattofit.exercise.application.sports.dto.SportsCreateRequest;
-import com.flab.eattofit.exercise.domain.sports.Sports;
-import com.flab.eattofit.exercise.domain.sports.SportsRepository;
-import com.flab.eattofit.exercise.exception.sports.exceptions.SportsAlreadyExistedException;
-import com.flab.eattofit.exercise.infrastructure.sports.SportsFakeRepository;
+import com.flab.eattofit.exercise.application.sports.sports.dto.SportsCreateRequest;
+import com.flab.eattofit.exercise.domain.sports.sports.Sports;
+import com.flab.eattofit.exercise.domain.sports.sports.SportsRepository;
+import com.flab.eattofit.exercise.exception.sports.sports.exceptions.SportsAlreadyExistedException;
+import com.flab.eattofit.exercise.infrastructure.sports.sports.SportsFakeRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
