+++ src/test/java/com/flab/eattofit/helper/MockBeanInjection.java
@@ -4,10 +4,10 @@ import com.amazonaws.services.s3.AmazonS3;
 import com.fasterxml.jackson.databind.json.JsonMapper;
 import com.flab.eattofit.exercise.application.fitness.fitness.FitnessService;
 import com.flab.eattofit.exercise.application.fitness.memberfitness.MemberFitnessesService;
-import com.flab.eattofit.exercise.application.sports.SportsService;
+import com.flab.eattofit.exercise.application.sports.sports.SportsService;
 import com.flab.eattofit.exercise.domain.fitness.fitness.FitnessRepository;
 import com.flab.eattofit.exercise.domain.fitness.memberfitness.MemberFitnessesRepository;
-import com.flab.eattofit.exercise.domain.sports.SportsRepository;
+import com.flab.eattofit.exercise.domain.sports.sports.SportsRepository;
 import com.flab.eattofit.food.application.FoodService;
 import com.flab.eattofit.member.application.auth.AuthService;
 import com.flab.eattofit.member.application.auth.OAuthManager;
